+++ src/simmate/apps/badelf/core/__init__.py
@@ -1,14 +1,5 @@
 # -*- coding: utf-8 -*-
 
-try:
-    from pybader.interface import Bader
-except:
-    raise Exception(
-        "Most functions of the BadELF app require the pybader package."
-        "Install this with `conda install -c conda-forge pybader`"
-    )
-
-
 from .badelf import BadElfToolkit
 from .electride_finder import ElectrideFinder
 from .partitioning import PartitioningToolkit
