+++ src/simmate/apps/materials_project/workflows/population_analysis/bader.py
@@ -2,9 +2,7 @@
 
 from pathlib import Path
 
-from simmate.apps.bader.workflows import (  # PopulationAnalysis__Bader__CombineChgcars,
-    PopulationAnalysis__Bader__Bader,
-)
+from simmate.apps.bader.workflows import PopulationAnalysis__Bader__Bader
 from simmate.apps.materials_project.workflows.static_energy.matproj import (
     StaticEnergy__Vasp__Matproj,
 )
@@ -35,13 +33,6 @@ class PopulationAnalysis__VaspBader__BaderMatproj(Workflow):
             source=source,
             directory=prebader_dir,
         ).result()
-
-        # # Setup chargecars for the bader analysis and wait until complete
-        # chgcomb_dir = directory / PopulationAnalysis__Bader__CombineChgcars.name_full
-        # PopulationAnalysis__Bader__CombineChgcars.run(
-        #     directory=chgcomb_dir,
-        #     previous_directory=prebader_dir,
-        # ).result()
 
         # And run the bader analysis on the resulting chg denisty
         bader_dir = directory / PopulationAnalysis__Bader__Bader.name_full
